+++ hypervisor/arch/x86/seed/seed_abl.c
@@ -4,7 +4,10 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <hypervisor.h>
+#include <types.h>
+#include <cpu.h>
+#include <pgtable.h>
+#include <rtl.h>
 #include <seed.h>
 #include "seed_abl.h"
